pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "py-auto-login"
version = "1.0.0"
description = "基于 Playwright 和 Tesseract OCR 的自动登录工具"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Auto Login Script"}
]
keywords = ["playwright", "tesseract", "ocr", "automation", "login"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
]

dependencies = [
    "playwright>=1.40.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "psutil>=5.9.0",
]

[project.scripts]
auto-login = "main:main"

[tool.setuptools]
packages = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
